pyproject.toml
[project]
name = "aws-lambda-opentelemetry"
version = "0.0.0"
description = "AWS Lambda OpenTelemetry tools for Python"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["opentelemetry", "lambda", "aws", "aws lambda", "faas"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Rust",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
]
dependencies = [
    "opentelemetry-api>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "C4",
    "UP",
]
[tool.ruff.lint.pycodestyle]
max-line-length = 120
